Rework reading spring web headers
