+++ instrumentation/spring/spring-web/spring-web-3.1/library/src/main/java/io/opentelemetry/instrumentation/spring/web/v3_1/SpringWebHttpAttributesGetter.java
@@ -5,15 +5,12 @@
 
 package io.opentelemetry.instrumentation.spring.web.v3_1;
 
-import static java.util.Collections.emptyList;
-
 import io.opentelemetry.instrumentation.api.semconv.http.HttpClientAttributesGetter;
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodType;
 import java.util.List;
 import javax.annotation.Nullable;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpRequest;
 import org.springframework.http.client.ClientHttpResponse;
 
@@ -23,7 +20,6 @@ enum SpringWebHttpAttributesGetter
 
   @Nullable private static final MethodHandle GET_STATUS_CODE;
   @Nullable private static final MethodHandle STATUS_CODE_VALUE;
-  @Nullable private static final MethodHandle GET_HEADERS;
 
   static {
     MethodHandles.Lookup lookup = MethodHandles.publicLookup();
@@ -58,24 +54,6 @@ enum SpringWebHttpAttributesGetter
 
     GET_STATUS_CODE = getStatusCode;
     STATUS_CODE_VALUE = statusCodeValue;
-
-    // since spring web 7.0
-    MethodHandle methodHandle =
-        findGetHeadersMethod(MethodType.methodType(List.class, String.class, List.class));
-    if (methodHandle == null) {
-      // up to spring web 7.0
-      methodHandle =
-          findGetHeadersMethod(MethodType.methodType(Object.class, Object.class, Object.class));
-    }
-    GET_HEADERS = methodHandle;
-  }
-
-  private static MethodHandle findGetHeadersMethod(MethodType methodType) {
-    try {
-      return MethodHandles.lookup().findVirtual(HttpHeaders.class, "getOrDefault", methodType);
-    } catch (Throwable t) {
-      return null;
-    }
   }
 
   @Override
@@ -91,19 +69,7 @@ enum SpringWebHttpAttributesGetter
 
   @Override
   public List<String> getHttpRequestHeader(HttpRequest httpRequest, String name) {
-    return getHeader(httpRequest.getHeaders(), name);
-  }
-
-  @SuppressWarnings("unchecked") // casting MethodHandle.invoke result
-  private static List<String> getHeader(HttpHeaders headers, String name) {
-    if (GET_HEADERS != null) {
-      try {
-        return (List<String>) GET_HEADERS.invoke(headers, name, emptyList());
-      } catch (Throwable t) {
-        // ignore
-      }
-    }
-    return emptyList();
+    return HeaderUtil.getHeader(httpRequest.getHeaders(), name);
   }
 
   @Override
@@ -125,7 +91,7 @@ enum SpringWebHttpAttributesGetter
   @Override
   public List<String> getHttpResponseHeader(
       HttpRequest httpRequest, ClientHttpResponse clientHttpResponse, String name) {
-    return getHeader(clientHttpResponse.getHeaders(), name);
+    return HeaderUtil.getHeader(clientHttpResponse.getHeaders(), name);
   }
 
   @Override
